Move `PatientPrefillInfo` to `patient` package
