+++ app/src/main/java/org/simple/clinic/instantsearch/InstantSearchScreenKey.kt
@@ -5,7 +5,7 @@ import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.parcelize.Parcelize
 import org.simple.clinic.navigation.v2.ScreenKey
 import org.simple.clinic.patient.businessid.Identifier
-import org.simple.clinic.scanid.PatientPrefillInfo
+import org.simple.clinic.patient.PatientPrefillInfo
 
 @Parcelize
 data class InstantSearchScreenKey(
