+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdEvent.kt
@@ -1,6 +1,7 @@
 package org.simple.clinic.scanid
 
 import org.simple.clinic.patient.Patient
+import org.simple.clinic.patient.PatientPrefillInfo
 import org.simple.clinic.patient.businessid.Identifier
 import org.simple.clinic.widgets.UiEvent
