+++ app/src/main/java/org/simple/clinic/scanid/ScanSimpleIdModel.kt
@@ -2,6 +2,7 @@ package org.simple.clinic.scanid
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
+import org.simple.clinic.patient.PatientPrefillInfo
 import org.simple.clinic.scanid.ScanSearchState.NotSearching
 import org.simple.clinic.scanid.ScanSearchState.Searching
